backend/pyproject.toml
[project]
name = "quiz-backend"
version = "0.1.0"
description = "FastAPI backend for Quiz Taker ChatKit app"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.114.1,<0.116",
    "uvicorn[standard]>=0.36,<0.37",
    "openai>=1.40",
    "openai-agents>=0.0.3",
    "openai-chatkit>=1.4.0,<2",
    "jinja2",
]

[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

